requirements.txt
asgiref==3.8.1
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.2.1
reportlab==4.4.2
sqlparse==0.5.3
tzdata==2025.2
